Changes: - changed C# client to use interfaces to generate the JObject to send in the request. - changed position to real coordinates and altitude to double. - added "speed" to the API and database - changed latitude, longitude and altitude fields in the database from int to float

+++ client/Client/Sensors/AltitudeSensorInterface.cs
@@ -7,6 +7,6 @@ namespace Client.Sensors
 {
     internal interface AltitudeSensorInterface
     {
-        int GetAltitude();
+        double GetAltitude();
     }
 }
